LF- 14699: drop indirection layers (#2127)
- [x] Implements modular refactor
- [x] fixes Instruction types (We pass array of arrays)

We added a modular layer for zaps. This PR drops the genericity layer (ZapperStrategy) that is not useful anymore, and creates too many level of indirections.
